Fix elections
